refactor: 컴포넌트 이름을 수정한다.
